% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Computes the values}
\usage{
pv(FV, r, n = 5)
}
\arguments{
\item{FV}{\code{numeric} The future value of the thing}

\item{r}{\code{numeric} vector of legth l for the rate}

\item{n}{the number of years that the rate \code{r} is applied}
}
\description{
This is the description of the function that says the function computes an output value for every input
}
\examples{
\dontrun{
pv(FV = 1000, r = .08, n = 5)
[1] 680.58
pv(1000, .08, 5)
[1] 680.58

pv(r = .08, FV = 1000, n = 5)
[1] 680.58
}
}
