% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexis_function.R
\name{rexicute}
\alias{rexicute}
\title{rexicute}
\usage{
rexicute(datafile_string)
}
\description{
This function allows user to employ Rexis from the R console.
}
\examples{
rexicute()
   A new window will allow user to select a .csv file. Navigate to Rexis/data/heart.csv
}
