% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexis_function.R
\name{rexicuteshiny}
\alias{rexicuteshiny}
\title{rexicute Shiny}
\usage{
rexicuteshiny(datafile_string)
}
\arguments{
\item{datafile_string}{this is a text example}
}
\description{
this is the sample description for rexicute
}
\examples{
rexicute ('C:/Users/mchale/OneDrive/Documents/AFIT/WInter 20/OPER 782 Data Science Programs/Rexis/data/heart.csv')
}
