% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexis_function.R
\name{rexicuteshiny}
\alias{rexicuteshiny}
\title{Rexicute for Shiny}
\usage{
rexicuteshiny(datafile_string)
}
\arguments{
\item{()}{there are no inputs to this function. An action button in the app will prompt user for a .csv file.}
}
\description{
This R function provides a framework to launch the Rexis R Shiny app.
}
\examples{
rexicute()
   click "browse" and navigate to Rexis/data/heart.csv
}
